+++ nbforager/nb_forager.py
@@ -301,7 +301,7 @@ class NbForager:
             - ``_power_outlets``
             - ``_power_ports``
             - ``_rear_ports``
-            - ``_virtual_chassis_members``
+            - ``_vc_members``
 
             In dcim.interfaces, virtualization.interfaces: